+++ examples/web/wsgi_noframework.py
@@ -37,7 +37,7 @@ def application(environ, start_response):
 
 engine = create_engine("mysql+pymysql://root:secret@localhost/testdb")
 data_store = SQLAlchemyDataStore(engine)
-event_broker = RedisEventBroker.from_url("redis://localhost")
+event_broker = RedisEventBroker("redis://localhost")
 scheduler = Scheduler(data_store, event_broker)
 scheduler.add_schedule(tick, IntervalTrigger(seconds=1), id="tick")
 scheduler.start_in_background()