+++ src/apscheduler/eventbrokers/redis.py
@@ -3,12 +3,12 @@ from __future__ import annotations
 from asyncio import CancelledError
 from contextlib import AsyncExitStack
 from logging import Logger
-from typing import Any
 
 import anyio
 import attrs
 import tenacity
 from anyio import move_on_after
+from attr.validators import instance_of
 from redis import ConnectionError
 from redis.asyncio import Redis
 from redis.asyncio.client import PubSub
@@ -27,32 +27,33 @@ class RedisEventBroker(BaseExternalEventBroker):
 
     .. _redis: https://pypi.org/project/redis/
 
-    :param client: an asynchronous Redis client
+    :param client_or_url: an asynchronous Redis client or a Redis URL
+        (```redis://...```)
     :param channel: channel on which to send the messages
     :param stop_check_interval: interval (in seconds) on which the channel listener
         should check if it should stop (higher values mean slower reaction time but less
         CPU use)
+
+    .. note:: The event broker will not manage the life cycle of any client instance
+        passed to it, so you need to close the client afterwards when you're done with
+        it.
     """
 
-    client: Redis
+    client_or_url: Redis | str = attrs.field(validator=instance_of((Redis, str)))
     channel: str = attrs.field(kw_only=True, default="apscheduler")
     stop_check_interval: float = attrs.field(kw_only=True, default=1)
-    _close_client_on_exit: bool = attrs.field(kw_only=True, default=False)
+
+    _client: Redis = attrs.field(init=False)
+    _close_on_exit: bool = attrs.field(init=False, default=False)
     _stopped: bool = attrs.field(init=False, default=True)
 
-    @classmethod
-    def from_url(cls, url: str, **kwargs: Any) -> RedisEventBroker:
-        """
-        Create a new event broker from a URL.
-
-        :param url: a Redis URL (```redis://...```)
-        :param kwargs: keyword arguments to pass to the initializer of this class
-        :return: the newly created event broker
-
-        """
-        pool = ConnectionPool.from_url(url)
-        client = Redis(connection_pool=pool)
-        return cls(client, close_client_on_exit=True, **kwargs)
+    def __attrs_post_init__(self) -> None:
+        if isinstance(self.client_or_url, str):
+            pool = ConnectionPool.from_url(self.client_or_url)
+            self._client = Redis(connection_pool=pool)
+            self._close_on_exit = True
+        else:
+            self._client = self.client_or_url
 
     def _retry(self) -> tenacity.AsyncRetrying:
         def after_attempt(retry_state: tenacity.RetryCallState) -> None:
@@ -72,17 +73,17 @@ class RedisEventBroker(BaseExternalEventBroker):
             reraise=True,
         )
 
-    async def _disconnect(self) -> None:
+    async def _close_client(self) -> None:
         with move_on_after(5, shield=True):
-            await self.client.aclose(close_connection_pool=True)
+            await self._client.aclose(close_connection_pool=True)
 
     async def start(self, exit_stack: AsyncExitStack, logger: Logger) -> None:
         # Close the client and its connection pool if this broker was created using
         # .from_url()
-        if self._close_client_on_exit:
-            exit_stack.push_async_callback(self._disconnect)
+        if self._close_on_exit:
+            exit_stack.push_async_callback(self._close_client)
 
-        pubsub = await exit_stack.enter_async_context(self.client.pubsub())
+        pubsub = await exit_stack.enter_async_context(self._client.pubsub())
         await pubsub.subscribe(self.channel)
         await super().start(exit_stack, logger)
 
@@ -120,4 +121,4 @@ class RedisEventBroker(BaseExternalEventBroker):
         notification = self.generate_notification(event)
         async for attempt in self._retry():
             with attempt:
-                await self.client.publish(self.channel, notification)
+                await self._client.publish(self.channel, notification)